+++ src/background/runtime/mount-runtime-listener.ts
@@ -1,12 +1,10 @@
 import { browser } from '../../service/browser';
 import type { BgReqAction } from '../../service/runtime/background';
 import { handleBgReqAction } from '../../service/runtime/background';
-import { detectLanguage } from '../../service/tabs/detect-language';
 import { convertClipboard } from '../clipboard';
 import { getConverter } from '../converter';
 import { bgLog } from '../logger';
 import { bgGetPref } from '../state/storage';
-import { getTargetByAutoConvert } from './handle-get-auto-convert';
 import { getTargetByFilter } from './handle-get-filter-target';
 import { getTarget } from './handle-get-target';
 
@@ -20,14 +18,10 @@ export function mountRuntimeListener() {
 
     return bgGetPref().then(async pref => {
       switch (action.type) {
-        case 'AutoConvert':
-          return handleBgReqAction(action, getTargetByAutoConvert(sender.tab!.id!));
         case 'FilterTarget':
           return handleBgReqAction(action, getTargetByFilter(pref, sender.url!));
         case 'GetTarget':
-          return handleBgReqAction(action, getTarget(pref, sender));
-        case 'DetectLang':
-          return handleBgReqAction(action, detectLanguage(sender.tab!.id));
+          return handleBgReqAction(action, getTarget(pref, sender, action.payload.zhType));
         case 'NodesText':
           return getConverter().then(async converter =>
             handleBgReqAction(